+++ package.json
@@ -25,6 +25,7 @@
     "@slack/rtm-api": "^6.1.0",
     "@slack/web-api": "^6.8.1",
     "dotenv": "^16.0.3",
-    "googleapis": "^118.0.0"
+    "googleapis": "^118.0.0",
+    "luxon": "^3.3.0"
   }
 }
