+++ src/index.js
@@ -1,73 +1,11 @@
 require("dotenv").config();
 const { App } = require("@slack/bolt");
-
-const fs = require("fs").promises;
-const path = require("path");
-const { authenticate } = require("@google-cloud/local-auth");
 const { google } = require("googleapis");
+const { DateTime } = require('luxon');
+
 const extractDate = require("./date-extract");
+const authorize = require("./auth");
 
-// If modifying these scopes, delete token.json.
-const SCOPES = ["https://www.googleapis.com/auth/calendar", "https://www.googleapis.com/auth/calendar.events"];
-
-// The file token.json stores the user's access and refresh tokens, and is
-// created automatically when the authorization flow completes for the first
-// time.
-const TOKEN_PATH = path.join(process.cwd(), "token.json");
-const CREDENTIALS_PATH = path.join(process.cwd(), "credentials.json");
-
-/**
- * Reads previously authorized credentials from the save file.
- *
- * @return {Promise<OAuth2Client|null>}
- */
-async function loadSavedCredentialsIfExist() {
-  try {
-    const content = await fs.readFile(TOKEN_PATH);
-    const credentials = JSON.parse(content);
-    return google.auth.fromJSON(credentials);
-  } catch (err) {
-    return null;
-  }
-}
-
-/**
- * Serializes credentials to a file compatible with GoogleAUth.fromJSON.
- *
- * @param {OAuth2Client} client
- * @return {Promise<void>}
- */
-async function saveCredentials(client) {
-  const content = await fs.readFile(CREDENTIALS_PATH);
-  const keys = JSON.parse(content);
-  const key = keys.installed || keys.web;
-  const payload = JSON.stringify({
-    type: "authorized_user",
-    client_id: key.client_id,
-    client_secret: key.client_secret,
-    refresh_token: client.credentials.refresh_token,
-  });
-  await fs.writeFile(TOKEN_PATH, payload);
-}
-
-/**
- * Load or request or authorization to call APIs.
- *
- */
-async function authorize() {
-  let client = await loadSavedCredentialsIfExist();
-  if (client) {
-    return client;
-  }
-  client = await authenticate({
-    scopes: SCOPES,
-    keyfilePath: CREDENTIALS_PATH,
-  });
-  if (client.credentials) {
-    await saveCredentials(client);
-  }
-  return client;
-}
 
 /**
  * Lists the next 10 events on the user's primary calendar.
@@ -113,19 +51,19 @@ async function handleScheduling(message) {
   }
 
   // TODO: Schedule event.
-  return date;
+
+
+  return date.toLocaleString(DateTime.DATETIME_FULL);
 }
 
 app.message(async ({ message, say }) => {
   try {
-    const result = await handleScheduling(message.text);
-    if (result !== false) {
-      // respond
-      console.log("responding");
-      say(`I'm on it!  I'll schedule that for you at ${result}`);
+    const msg = await handleScheduling(message.text);
+    if (msg !== false) {
+      say(`I'm on it!  I'll schedule that for you at ${msg}`);
     }
-    console.log("responded, maybe");
   } catch (e) {
+    say('Something when wrong when making calendar.')
     console.log('COULDN"T Send msag', e);
   }
 });
