+++ src/test.ts
@@ -1,0 +1,3 @@
+const extractDate = require('./date-extract');
+
+extractDate('on wed 4/19 at 9:30am');
